Split off multiple sheet unflattening tests into their own file
